# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bakery Service API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Daftar Produk",
        "description": "Mengambil daftar seluruh produk beserta informasi harga dan stok yang tersedia. Dapat difilter berdasarkan kategori produk.",
        "operationId": "list_products_v1_products_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter berdasarkan kategori produk (misal: 'bread', 'pastry')",
              "examples": [
                "bread"
              ],
              "title": "Category"
            },
            "description": "Filter berdasarkan kategori produk (misal: 'bread', 'pastry')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/products/{product_id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Detail Produk",
        "description": "Mengambil rincian informasi produk berdasarkan ID produk.",
        "operationId": "get_product_v1_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik produk",
              "examples": [
                "prod_sourdough01"
              ],
              "title": "Product Id"
            },
            "description": "ID unik produk"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/products/{product_id}/stock": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Perbarui Stok Produk [Idempoten - HTTP PUT]",
        "description": "Menetapkan nilai stok absolut untuk suatu produk roti. Operasi HTTP PUT ini bersifat **Idempoten** karena jika dipanggil berulang kali dengan nilai stok yang sama, nilai akhir stok pada sistem tetap sama.",
        "operationId": "update_product_stock_v1_products__product_id__stock_put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik produk",
              "examples": [
                "prod_sourdough01"
              ],
              "title": "Product Id"
            },
            "description": "ID unik produk"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductStockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/payments": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Memproses Pembayaran [Idempoten - Pola Stripe]",
        "description": "**[Operasi Idempoten - Pola Stripe Finansial]**\n\nMemproses transaksi pembayaran pesanan toko roti. Operasi ini dilindungi oleh header `Idempotency-Key` (UUID v4) guna menjamin bahwa pelanggan tidak akan terdebit dua kali (*double charging*) saat terjadi kegagalan/timeout jaringan. Pengulangan permintaan dengan kunci sama akan mengembalikan respons *cache* (dengan header `Idempotent-Replay: true` dan `X-Cache-Lookup: HIT`).",
        "operationId": "create_payment_v1_payments_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik UUID v4 untuk mencegah transaksi pembayaran ganda",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik UUID v4 untuk mencegah transaksi pembayaran ganda"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pembayaran berhasil diproses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Nominal pembayaran tidak sesuai atau status pesanan tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Konflik kunci idempotensi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments/{payment_id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Detail Pembayaran",
        "description": "Mengambil informasi dan status transaksi pembayaran berdasarkan ID pembayaran.",
        "operationId": "get_payment_v1_payments__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik pembayaran",
              "examples": [
                "pay_trx_001"
              ],
              "title": "Payment Id"
            },
            "description": "ID unik pembayaran"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pembayaran tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments/{payment_id}/confirm": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Konfirmasi Pembayaran [Idempoten - State Transition]",
        "description": "**[Operasi Idempoten - State Transition]**\n\nMelakukan konfirmasi status penyelesaian transaksi pembayaran. Operasi ini idempoten karena menetapkan status akhir ke `settlement`. Pemanggilan berulang tidak mengubah hasil akhir.",
        "operationId": "confirm_payment_v1_payments__payment_id__confirm_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID unik pembayaran",
              "examples": [
                "pay_trx_001"
              ],
              "title": "Payment Id"
            },
            "description": "ID unik pembayaran"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kunci unik UUID v4",
              "title": "Idempotency-Key"
            },
            "description": "Kunci unik UUID v4"
          }
        ],
        "responses": {
          "200": {
            "description": "Pembayaran dikonfirmasi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pembayaran tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFCErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "invalid_request_error"
            ]
          },
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "idempotency_conflict"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Keys for idempotent requests cannot be reused with different request body."
            ]
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param",
            "examples": [
              "Idempotency-Key"
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentCreate": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id",
            "examples": [
              "ord_20260925_001"
            ]
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "examples": [
              90000
            ]
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method",
            "description": "Metode pembayaran: qris, va_bca, credit_card, cash.",
            "examples": [
              "qris"
            ]
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "amount",
          "payment_method"
        ],
        "title": "PaymentCreate"
      },
      "PaymentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "pay_trx_99a1b2"
            ]
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "payment",
            "examples": [
              "payment"
            ]
          },
          "order_id": {
            "type": "string",
            "title": "Order Id",
            "examples": [
              "ord_20260925_001"
            ]
          },
          "amount": {
            "type": "integer",
            "title": "Amount",
            "examples": [
              90000
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "idr",
            "examples": [
              "idr"
            ]
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method",
            "examples": [
              "qris"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status pembayaran: pending, settlement, denied, expired.",
            "examples": [
              "settlement"
            ]
          },
          "created": {
            "type": "integer",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "amount",
          "payment_method",
          "status"
        ],
        "title": "PaymentResponse"
      },
      "ProductListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list",
            "examples": [
              "list"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/products",
            "examples": [
              "/v1/products"
            ]
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "examples": [
              false
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "has_more",
          "data"
        ],
        "title": "ProductListResponse"
      },
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "prod_sourdough01"
            ]
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "product",
            "examples": [
              "product"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Artisan Sourdough Country Loaf"
            ]
          },
          "category": {
            "type": "string",
            "title": "Category",
            "examples": [
              "bread"
            ]
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price",
            "description": "Harga dalam satuan IDR.",
            "examples": [
              45000
            ]
          },
          "stock": {
            "type": "integer",
            "title": "Stock",
            "description": "Jumlah ketersediaan stok.",
            "examples": [
              20
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Roti sourdough fermentasi 24 jam alami tanpa ragi instan."
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "unit_price",
          "stock"
        ],
        "title": "ProductResponse"
      },
      "ProductStockUpdate": {
        "properties": {
          "stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock",
            "description": "Jumlah stok absolut yang ditetapkan",
            "examples": [
              50
            ]
          }
        },
        "type": "object",
        "required": [
          "stock"
        ],
        "title": "ProductStockUpdate"
      },
      "RFCErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "RFCErrorResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}